feat: Add interactive help modal for Facebook token
Adds a help icon and a detailed modal to guide users in generating a Facebook System User Token with the correct permissions. This clarifies the two-step process of creating the token (ads_read, pages_read_engagement) and then assigning the ad account asset with the 'Ver rendimiento' role. Also fixes the JS loading by adding the 'scripts' block to base.html.